Move TableParser to isatab_full module
